Update: Add notification features in NavNotificationPage
