+++ lib/Mobile/Notifications.dart
@@ -1,6 +1,5 @@
 import 'package:flutter/material.dart';
-import 'package:foodtracker_firebase/Properties/dashboardAssets/ImageSlider.dart';
-import 'package:foodtracker_firebase/Properties/dashboardAssets/foodDescription.dart';
+import 'package:foodtracker_firebase/Properties/notificationAssets/notificationBox.dart';
 
 class NavNotificationsPage extends StatefulWidget {
   const NavNotificationsPage({super.key});
@@ -13,129 +12,31 @@ class _NavNotificationsPageState extends State<NavNotificationsPage> {
   @override
   Widget build(BuildContext context) {
     return Scaffold(
-      body: Container(
-        width: double.infinity,
-        height: MediaQuery.of(context).size.height,
-        decoration: const BoxDecoration(
-          gradient: LinearGradient(
-            colors: [Color(0xff213448), Color(0xff213448)],
-            begin: Alignment.topCenter,
-            end: Alignment.bottomCenter,
+      appBar: AppBar(
+        automaticallyImplyLeading: false,
+        elevation: 0,
+        toolbarHeight: 100,
+        backgroundColor: const Color(0xff213448),
+        title: const Padding(
+          padding: EdgeInsets.symmetric(horizontal: 20.0, vertical: 10.0),
+          child: Text(
+            "Notifications",
+            style: TextStyle(
+              fontFamily: 'Montserrat',
+              color: Colors.white,
+              fontSize: 30,
+            ),
           ),
         ),
-        child: ListView(
-          padding: const EdgeInsets.symmetric(vertical: 20, horizontal: 16),
-          children: [
-            const ImageSlider(),
-            const SizedBox(height: 20),
-            const Text(
-              'Recommended Meals',
-              style: TextStyle(
-                fontFamily: 'Montserrat',
-                fontSize: 20,
-                fontWeight: FontWeight.w600,
-                color: Colors.white,
-              ),
-            ),
-            const SizedBox(height: 10),
+      ),
 
-            // 🔹 Meal 1
-            Card(
-              color: const Color(0xff2f4a5d),
-              margin: const EdgeInsets.symmetric(vertical: 8),
-              shape: RoundedRectangleBorder(
-                borderRadius: BorderRadius.circular(12),
-              ),
-              child: ListTile(
-                leading: ClipRRect(
-                  borderRadius: BorderRadius.circular(8),
-                  child: Image.asset(
-                    'images/food1.jpg',
-                    width: 50,
-                    height: 50,
-                    fit: BoxFit.cover,
-                  ),
-                ),
-                title: const Text(
-                  'Grilled Chicken Salad',
-                  style: TextStyle(
-                    color: Colors.white,
-                    fontWeight: FontWeight.bold,
-                  ),
-                ),
-                subtitle: const Text(
-                  'Fresh greens with protein',
-                  style: TextStyle(color: Colors.white70),
-                ),
-                // only the icon pushes the next page
-                trailing: IconButton(
-                  icon: const Icon(
-                    Icons.arrow_forward_ios,
-                    color: Colors.white54,
-                    size: 16,
-                  ),
-                  onPressed: () {
-                    Navigator.push(
-                      context,
-                      MaterialPageRoute(
-                        builder: (context) => Fooddescription(),
-                      ),
-                    );
-                  },
-                ),
-                // no onTap here: tapping card DOES NOTHING
-              ),
-            ),
-
-            // 🔹 Meal 2
-            Card(
-              color: const Color(0xff2f4a5d),
-              margin: const EdgeInsets.symmetric(vertical: 8),
-              shape: RoundedRectangleBorder(
-                borderRadius: BorderRadius.circular(12),
-              ),
-              child: ListTile(
-                leading: ClipRRect(
-                  borderRadius: BorderRadius.circular(8),
-                  child: Image.asset(
-                    'images/food2.jpg',
-                    width: 50,
-                    height: 50,
-                    fit: BoxFit.cover,
-                  ),
-                ),
-                title: const Text(
-                  'Avocado Toast',
-                  style: TextStyle(
-                    color: Colors.white,
-                    fontWeight: FontWeight.bold,
-                  ),
-                ),
-                subtitle: const Text(
-                  'Simple and nutritious',
-                  style: TextStyle(color: Colors.white70),
-                ),
-                trailing: IconButton(
-                  icon: const Icon(
-                    Icons.arrow_forward_ios,
-                    color: Colors.white54,
-                    size: 16,
-                  ),
-                  onPressed: () {
-                    Navigator.push(
-                      context,
-                      MaterialPageRoute(
-                        builder: (context) => Fooddescription(),
-                      ),
-                    );
-                  },
-                ),
-              ),
-            ),
-
-            // 🔹 Add more manually below...
-          ],
-        ),
+      // 👉 Just show one NotificationBox for now
+      body: ListView(
+        children: const [
+          NotificationBox(),
+          NotificationBox(),
+          NotificationBox(),
+        ],
       ),
     );
   }
