--- lib/Properties/trendingAsset/review_modal.dart
@@ -1,134 +1,0 @@
-import 'package:flutter/material.dart';
-import 'package:flutter_rating_bar/flutter_rating_bar.dart';
-
-class ReviewModal extends StatefulWidget {
-  final String restaurantName;
-  final Function(double rating, String comment) onSubmit;
-
-  const ReviewModal({
-    super.key,
-    required this.restaurantName,
-    required this.onSubmit,
-  });
-
-  @override
-  State<ReviewModal> createState() => _ReviewModalState();
-}
-
-class _ReviewModalState extends State<ReviewModal> {
-  final TextEditingController commentController = TextEditingController();
-  double rating = 0;
-
-  @override
-  Widget build(BuildContext context) {
-    return DraggableScrollableSheet(
-      expand: false,
-      initialChildSize: 0.65,
-      minChildSize: 0.4,
-      maxChildSize: 0.9,
-      builder: (context, scrollController) {
-        return Container(
-          padding: const EdgeInsets.all(20),
-          decoration: const BoxDecoration(
-            color: Colors.white,
-            borderRadius: BorderRadius.vertical(top: Radius.circular(24)),
-          ),
-          child: Column(
-            crossAxisAlignment: CrossAxisAlignment.center,
-            children: [
-              // Header
-              Row(
-                mainAxisAlignment: MainAxisAlignment.spaceBetween,
-                children: [
-                  const SizedBox(width: 40),
-                  const Text(
-                    "Leave a Review",
-                    style: TextStyle(
-                      fontFamily: 'Montserrat',
-                      fontSize: 20,
-                      fontWeight: FontWeight.bold,
-                    ),
-                  ),
-                  IconButton(
-                    icon: const Icon(Icons.close, color: Colors.black54),
-                    onPressed: () => Navigator.pop(context),
-                  ),
-                ],
-              ),
-              const SizedBox(height: 10),
-              Text(
-                "Rate your experience with ${widget.restaurantName}",
-                style: const TextStyle(
-                  fontFamily: 'Montserrat',
-                  fontSize: 14,
-                  color: Colors.grey,
-                ),
-              ),
-              const SizedBox(height: 20),
-
-              // Rating Bar
-              RatingBar.builder(
-                initialRating: 0,
-                minRating: 1,
-                allowHalfRating: true,
-                itemCount: 5,
-                itemSize: 40,
-                unratedColor: Colors.grey.shade300,
-                itemBuilder: (context, _) =>
-                    const Icon(Icons.star_rounded, color: Colors.amber),
-                onRatingUpdate: (value) {
-                  setState(() => rating = value);
-                },
-              ),
-              const SizedBox(height: 30),
-
-              // Comment
-              TextField(
-                controller: commentController,
-                maxLines: 4,
-                decoration: InputDecoration(
-                  hintText: "What's on your mind?",
-                  filled: true,
-                  fillColor: const Color(0xfff7f7f7),
-                  border: OutlineInputBorder(
-                    borderRadius: BorderRadius.circular(16),
-                    borderSide: BorderSide.none,
-                  ),
-                ),
-              ),
-              const Spacer(),
-
-              // Submit
-              SizedBox(
-                width: double.infinity,
-                height: 55,
-                child: ElevatedButton(
-                  style: ElevatedButton.styleFrom(
-                    backgroundColor: const Color(0xFF547792),
-                    shape: RoundedRectangleBorder(
-                      borderRadius: BorderRadius.circular(16),
-                    ),
-                    elevation: 2,
-                  ),
-                  onPressed: () {
-                    widget.onSubmit(rating, commentController.text);
-                    Navigator.pop(context);
-                  },
-                  child: const Text(
-                    "Submit Review",
-                    style: TextStyle(
-                      fontFamily: 'Montserrat',
-                      fontSize: 16,
-                      color: Colors.white,
-                      fontWeight: FontWeight.w600,
-                    ),
-                  ),
-                ),
-              ),
-            ],
-          ),
-        );
-      },
-    );
-  }
-}
