+++ pubspec.yaml
@@ -51,7 +51,6 @@ dependencies:
   cloud_firestore: ^6.0.1
   flutter_rating_bar: ^4.0.1
 
-
 dev_dependencies:
   flutter_test:
     sdk: flutter
@@ -77,7 +76,6 @@ flutter:
   # To add assets to your application, add an assets section, like this:
   assets:
     - images/
-    - images/foodtracker.jpg
     - images/foodtracker_dark.png
 
   # An image asset can refer to one or more resolution-specific "variants", see
